Make GrowingPoolAllocator constructor take size_t args.
Also make growSize_ & chunkHeaderSize_ size_t, no reason they should not be since they are only used in conjunction with other vars that are size_t.

+++ Pulp/Common/Memory/AllocationPolicies/GrowingPoolAllocator.cpp
@@ -25,8 +25,8 @@ namespace {
 }
 
 
-GrowingPoolAllocator::GrowingPoolAllocator(unsigned int maxSizeInBytes, unsigned int growSize, 
-	unsigned int chunkHeaderSize, size_t maxElementSize, size_t maxAlignment, size_t offset)
+GrowingPoolAllocator::GrowingPoolAllocator(size_t maxSizeInBytes, size_t growSize,
+	size_t chunkHeaderSize, size_t maxElementSize, size_t maxAlignment, size_t offset)
 {
 	virtualStart_ = static_cast<char*>( VirtualMem::ReserveAddressSpace( maxSizeInBytes ) );
 	virtualEnd_ = &virtualStart_[ maxSizeInBytes ];
